Trying to come up with simple simulation. PS: broken commit

+++ fantoch/src/run/mod.rs
@@ -633,7 +633,8 @@ pub mod tests {
         // - and so on..
         let region_index = |process_id| {
             let mut index = process_id;
-            let n = config.n() as u8;
+            //let n = config.n() as u8;
+            let n = config.n() as u64;
             while index > n {
                 index -= n;
             }
